feat: enhance MindsDB integration and refactor Docker setup for improved application startup
