+++ self-host/.env.example
@@ -2,8 +2,8 @@
 # Required environment variables for the KBNet self-hosted instance
 DATABASE_URL=postgresql://mindsdb:mindsdb@kbnet_database:5432/mindsdb
 
-NEXT_PUBLIC_SERVER_URL=http://kbnet_server:8000
-NEXT_PUBLIC_WS_SERVER_URL=ws://kbnet_server:8000
+NEXT_PUBLIC_SERVER_URL=http://localhost:8000
+NEXT_PUBLIC_WS_SERVER_URL=ws://localhost:8000
 
 BETTER_AUTH_SECRET=auth_secret_token
 BETTER_AUTH_URL=http://localhost:3000
