+++ self-host/seed/src/lib/mindsdb.js
@@ -4,7 +4,7 @@ import dotenv from "dotenv";
 dotenv.config();
 
 const config = {
-  host: process.env.MINDSDB_HOST || "http://localhost:47334",
+  host: "http://localhost:47334",
   user: process.env.MINDSDB_USER || "mindsdb",
   password: process.env.MINDSDB_PASSWORD || "mindsdb",
 };
